move _open_file helper decorator from rg16 to core/util
