+++ obspy/io/rg16/util.py
@@ -5,33 +5,11 @@ from future.utils import native_str as nstr
 
 from codecs import encode
 import copy
-import decorator
 
 import numpy as np
 from obspy import Trace
-
-
-@decorator.decorator
-def _open_file(func, *args, **kwargs):
-    """
-    Ensure a file buffer is passed as first argument to the
-    decorated function.
-
-    :param func: callable that takes at least one argument;
-        the first argument must be treated as a buffer.
-    :return: callable
-    """
-    first_arg = args[0]
-    try:
-        with open(first_arg, 'rb') as fi:
-            args = tuple([fi] + list(args[1:]))
-            return func(*args, **kwargs)
-    except TypeError:  # assume we have been passed a buffer
-        if not hasattr(args[0], 'read'):
-            raise  # type error was in function call, not in opening file
-        out = func(*args, **kwargs)
-        first_arg.seek(0)  # reset position to start of file
-    return out
+# keep this import because _open_file used to be defined in here
+from obspy.core.util.decorator import _open_file  # NOQA
 
 
 def _read(fi, position, length, dtype, left_part=True):
